Remove failures.py
I think the implementation could be made much simpler if `Reader.feed_row()`
would just raise an exception if there were problems with the parsing. All of
`failures.py` could then be thrown away. Then, with the addition of a mere few
lines in the function `load_vicon_file`, the same effect could be achieved: that
function can keep track of the line number and reraise the exception with the
correct message, which is currently the job of `failures.Validator`.

Also, `FailableResult` and `FailableMixin` are there to monadically combine
intermediate parsing results. They were only useful because of the way the code
is currently designed. Suppose there were several different ways in which the
parsing of a single line could fail. Ideally, we'd want to present an
informative message to the user explaning how the process failed. Currently,
however, failures must be reported exactly once per line, by calling
`Validator.validate`. This constraint makes it necessary for states to keep
track of its different possible failure reasons and combine them somehow, all of
which is unnecessarily complex. Raising exceptions is a perfectly good way to
achieve the desired effect.