Add _UpdateStateMixin to simplify most state transitions
The only transition which isn't boringly simple is the one for
GettingMeasurementsState since it cares whether it's transitioning to a new
section or if the file ended. Either this mixin will be useful for all
_ReaderState subclasses except for it, or it'll be used for all of them in which
case the mixin should just be incorporated into _ReaderState itself.

I'll think about it later, for now the idea is just to supplant
_StepByStepReaderState with mixins.