Add ForcePlateGrouper stub
Idea is to factor out this functionality from ForcesEMGDevicesState. Main reason
why is to make this particular aspect of the latter's responsibilities more
easily unit testable. This obviously could be a function I decided to leave
everything in this module in classes to try to reason in a more OO way.