Update Sphinx cross-references to add ~ prefix.
The result is just more readable. For example,
`muscle_synergies.vicon_data.aggregator.Aggregator` becomes simply
`Aggregator`. In some cases, there is ambiguity in just keeping the name
of the class or method being referenced, but in the majority of the time
that version is simply better.

This is the link to the relevant docs:
https://www.sphinx-doc.org/en/master/usage/restructuredtext/roles.html#cross-referencing-syntax