Drop support for row indexing in DeviceData
`__getitem__` had an error since `DataFrame.iloc` does support receiving
a tuple (in which case it works like a 2D `ndarray`).
The method as it was could not distinguish between `(array_row,
array_col)` and `(frame, subframe)`.

Workarounds would be possible but the ones I could conceive were ugly.
This was unlikely to be needed so I threw it out.